Figuring out how to pass collections.
Forms seem to only handle static data that's always present. To use Option types
and variable amounts of data, which deserializes to something like a Vec, we
need to use web::Json<T> where T is the target of deserialization. This will
inflate the amount of models needed probably. Should reconsider how to structure
the views and models.